v0.53: fix boot hang — CRC read stored in size_t wrapped on -1
FitWriter::finish() read the file back for the CRC with the byte count in
a size_t; File::read() returns int and yields -1 on error, so a read
error made the loop count ~1.8e19 and spin forever (reading out of
bounds). Boot-time ride recovery calls finish() via repair(), so one bad
read during recovery hung setup() entirely — tasks never started and the
SD was never mounted, i.e. "device can't recognize the SD card" even
though the card is fine. Use a signed int and bail on <= 0.

+++ src/fit_writer.cpp
@@ -366,9 +366,12 @@ bool FitWriter::finish(time_t endUtc, double totalDistanceM, uint32_t timerS) {
     file_.seek(0);
     uint16_t crc = 0;
     uint8_t chunk[256];
-    size_t n;
+    // read() returns int and gives -1 on error; a size_t here would wrap to a
+    // huge value and spin forever (the boot-time recovery hit this and hung the
+    // whole device, so the SD never got mounted).
+    int n;
     while ((n = file_.read(chunk, sizeof(chunk))) > 0) {
-        for (size_t i = 0; i < n; ++i) crc = crc16(crc, chunk[i]);
+        for (int i = 0; i < n; ++i) crc = crc16(crc, chunk[i]);
     }
     file_.seek(12 + dataSize_);
     uint8_t crcBuf[2];
